% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Simulation of the queuing system times of a bank}
\format{
A data frame with 100 rows and 2 variables:
\describe{
\item{arrival_time}{arrival time of a customer, in seconds}
\item{service_time}{service time of a customer, in seconds}
}
}
\usage{
bank
}
\description{
A dataset containing the customer arrival time and service time
}
\keyword{datasets}
