% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Simulate the queuing system of a bank}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{a vector of customer arrival times}

\item{ServiceTimes}{a vector of customer service times}

\item{NumServers}{an integer of the number of service servers of a bank}
}
\value{
a data.frame simulating the queuing system of the bank:
\describe{
\item{Arrivals}{arrival time of a customer, in seconds}
\item{ServiceBegins}{the time that the customer service begins}
\item{ChosenServer}{the server that the customer goes through}
\item{ServiceEnds}{the time that the customer service ends}
}
}
\description{
This function takes information about customer arrival and service times
and simulates the queuing system of a bank.
}
\examples{
Multiserver(bank$arrival_time, bank$service_time, 3)
}
